% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalGetComponentKwds}
\alias{externalGetComponentKwds}
\title{Get FVS keywords for the components in a run.}
\usage{
externalGetComponentKwds(prjDir = getwd(), runUUID)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the 
current directory is the project directory.}

\item{runUUID}{a character vector of the run uuid that is processed}
}
\value{
a list of two data.frames corresponding to components attached
  to "Groups" and another to "Stands". Each data.frame has 4 columns as
  follows: 1-the group name or stand ID, 2-the component title, 3-the 
  component UUID, 4-the keywords.
}
\description{
Given a project directory a run uuid, get the keywords in the run.
}
