% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCallable.R
\name{externalMakeRun}
\alias{externalMakeRun}
\title{Build an FVS run in a project}
\usage{
externalMakeRun(
  prjDir = getwd(),
  title = NULL,
  standIDs = NULL,
  stdInit = "FVS_StandInit",
  variant,
  keywords = NULL,
  group = NULL
)
}
\arguments{
\item{prjDir}{is the path name to the project directory, if null the system
assumes that the current directory is the project directory.}

\item{title}{a character string with the run title, if null, the system generates 
the name, that is, it would be "Run 1", or "Run 2", and so on for 
all the runs already present. If none are present, then the default 
run title with be "Run 1" corresponding to the first run.}

\item{standIDs}{a vector of character strings holding the stand IDs (for example: 
\code{standIDs=c("id1","id2")} would load 2 stands, whereas: \code{standIDs="id1"} 
 would load just one. If NULL, all stands in the init table are loaded.}

\item{stdInit}{a character string of the name of the standinit table you 
want to use, the stands would be loaded from that table.}

\item{variant}{a 2 character string specifying the variant (required).
Ihe standID and the variant must match in the designated init table.}

\item{keywords}{a named vector of character strings with the keywords 
you want added to the group specified in the next argument 
(see below): Here is an example:
\code{keywords = c("title of first keyword set" = "Keyword line 1\\nKeyword line 2",
                    "title of the second set" = "Keyword line 1\\nKeyword line 2")}
All of the keywords will be set to be owned by the "base", that is, 
if you include "extension" keywords, they must already have the 
necessary extension start keyword (ie, for the FFE, its: FFIN) 
and the necessary End. Conditions, starting with IF must include "EndIf", no
checking is done. If the value of keywords is "NULL", then no keywords are added.}

\item{group}{a character string naming the group to which the keywords are attached. 
If NULL, then the keywords are attached to each stand.}
}
\value{
the uuid (identifier) of the new run
}
\description{
Build an FVS run in a project and add it to the list of runs in the project. 
If some FVS runs are already present, another is added with this call.
The name of the input data base is FVS_Data.db and it must already exist.
}
